updated admin panel: photos, genres CRUD, added login-logout

+++ admin/genres/delete.php
@@ -1,9 +1,15 @@
-<?php include "../db/admin_script.php"; ?>
 <?php
-connection();
+include "../../db/global_config.php";
+include "../db/admin_script.php";
+
 if (isset($_GET['g_id'])) {
     $the_genres_id = $_GET['g_id'];
-    deleteGenre($the_genres_id);
+
+    if (ifPhotoExists($the_genres_id)) {
+        echo "<script type='text/javascript'>alert('Some photos are under this genre already, cannot delete')</script>";
+    } else {
+        deleteGenre($the_genres_id);
+    }
     header("Location: index.php");
 }
 ?>
